Converted model<->H5 to model<->dict<->H5
